chore: optimize graphql from schema file
