+++ packages/core/codegen.yml
@@ -1,5 +1,5 @@
 overwrite: true
-schema: http://localhost:3000/api/graphql
+schema: ./@generated/graphql/schema.graphql
 documents:
   - ./src/**/*.{ts,tsx}
 config:
