+++ packages/core/package.json
@@ -4,8 +4,8 @@
   "license": "MIT",
   "browserslist": "supports es6-module and not dead",
   "scripts": {
-    "generate": "graphql-codegen",
-    "build": "yarn partytown && next build",
+    "generate": "tsx src/server/generateGraphqlSchemaFile.ts && graphql-codegen",
+    "build": "yarn partytown && tsx src/server/generateGraphqlSchemaFile.ts && next build",
     "dev": "yarn partytown && next dev",
     "clean": "rm -r .next",
     "serve": "next start",
@@ -37,6 +37,7 @@
     "@faststore/ui": "^2.1.78",
     "@graphql-tools/load-files": "^7.0.0",
     "@graphql-tools/merge": "^9.0.0",
+    "@graphql-tools/utils": "^10.0.5",
     "@types/react": "^18.0.14",
     "@vtex/client-cms": "^0.2.12",
     "autoprefixer": "^10.4.0",
@@ -61,6 +62,7 @@
     "style-loader": "^3.3.1",
     "swr": "^1.3.0",
     "tsconfig-paths-webpack-plugin": "^3.5.2",
+    "tsx": "^3.12.7",
     "typescript": "4.7.3"
   },
   "devDependencies": {
